Updated barcode view to common module
