+++ app/build.gradle
@@ -32,7 +32,6 @@ dependencies {
 
     implementation "com.google.android.gms:play-services-location:21.0.1"
 
-    implementation "com.google.zxing:core:3.4.1"
     implementation 'com.journeyapps:zxing-android-embedded:[4,5['
 
     implementation "com.google.android.play:review-ktx:[2,3["
