+++ app/src/main/java/movie/metropolis/app/screen/profile/component/MembershipCard.kt
@@ -1,39 +1,18 @@
 package movie.metropolis.app.screen.profile.component
 
-import android.graphics.Bitmap
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.LocalTextStyle
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.layout.onSizeChanged
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.unit.dp
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.*
+import androidx.compose.ui.draw.*
+import androidx.compose.ui.graphics.*
+import androidx.compose.ui.res.*
+import androidx.compose.ui.tooling.preview.*
+import androidx.compose.ui.unit.*
 import com.google.zxing.BarcodeFormat
-import com.google.zxing.MultiFormatWriter
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import movie.metropolis.app.R
+import movie.style.Barcode
 import movie.style.imagePlaceholder
 import movie.style.layout.StackedCardLayout
 import movie.style.textPlaceholder
@@ -129,34 +108,5 @@ private fun Preview() {
             until = "21.5.2053",
             points = "513.2"
         )
-    }
-}
-
-@Composable
-fun Barcode(
-    code: String,
-    format: BarcodeFormat,
-    modifier: Modifier = Modifier,
-    color: Color = Theme.color.container.background
-) {
-    val (size, onSizeChanged) = remember { mutableStateOf(IntSize.Zero) }
-    var image by remember { mutableStateOf(null as ImageBitmap?) }
-    LaunchedEffect(size, code, format) {
-        val writer = MultiFormatWriter()
-        image = withContext(Dispatchers.Default) {
-            val w = size.width
-            val h = size.height
-            if (w * h <= 0) return@withContext null
-            val matrix = writer.encode(code, format, w, h)
-            val bitmap = Bitmap.createBitmap(w, h, Bitmap.Config.ARGB_8888)
-            for (x in 0 until matrix.width) for (y in 0 until matrix.height) {
-                val pixel = if (matrix.get(x, y)) color.toArgb() else Color.Transparent.toArgb()
-                bitmap.setPixel(x, y, pixel)
-            }
-            bitmap.asImageBitmap()
-        }
-    }
-    Canvas(modifier.onSizeChanged(onSizeChanged)) {
-        drawImage(image ?: return@Canvas)
     }
 }
