+++ app/src/main/java/movie/metropolis/app/util/ActivityResult.kt
@@ -1,8 +1,5 @@
 package movie.metropolis.app.util
 
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
 import androidx.activity.result.ActivityResultRegistry
 import androidx.activity.result.contract.ActivityResultContract
 import kotlin.coroutines.resume
@@ -16,12 +13,4 @@ suspend fun <I, O> ActivityResultRegistry.register(
     register(key, contract) {
         cont.resume(it)
     }.launch(input)
-}
-
-fun Context.findActivity(): Activity {
-    when (this) {
-        is Activity -> return this
-        is ContextWrapper -> return baseContext.findActivity()
-    }
-    throw IllegalStateException("Unknown Context $this")
 }
