+++ style/build.gradle
@@ -36,6 +36,8 @@ dependencies {
     api "com.google.accompanist:accompanist-pager:0.29.1-alpha"
     api "com.google.accompanist:accompanist-placeholder-material:0.29.1-alpha"
 
+    api "com.google.zxing:core:3.5.1"
+
     implementation "androidx.startup:startup-runtime:[1,2["
 
 }
